+++ test/ayumi/plans/goal_progress_test.exs
@@ -1,5 +1,5 @@
 defmodule Ayumi.Plans.GoalProgressTest do
-  use Ayumi.DataCase, async: true
+  use Ayumi.DataCase, async: false
 
   alias Ayumi.Plans.GoalProgress